Move event delivery to a port
Introduce a new "port", backed by an iframe, over which
we send messages. The eventual goal is to move all
communication to ports, so that we can put each client
in a Web Worker.

In the future, I may want to use MessageChannel/MessagePort
directly... once that becomes better implemented, that is.

This causes the standard async redraw issues. I'll try to fix
it soon, I swear.