Don't error out on potential race conditions when moving the mouse
Since events and requests are sent asynchronously, it's possible to have
a ButtonPress and Motion queued before the client has processed them,
where the ButtonPress will drop the grab. We can't un-send protocol
events, and we don't want to take a sync grab, so just filter these out
on the client.

From my understanding, this is also the case in the real X11 case as
well, and clients just drop such events.