Clear pixels from damaged regions where we've copied
This fixes the close button appearing in the wrong color on window
frames.

When the frame is configured, we damage the entire frame, and then
configure the button. When we configure the button, we copy the
previously rendered button, and end up with no damage. Since we never
subtract the button's new region from the frame's damage, it ends up
painting over it. Solve this by clearing the damage region when we
copy the area.