server: Move the drawing model towards what X actually provides
X does not track outstanding damages across windows; it simply sends expose
events and lets X clients draw, clipped to their extents, siblings, and
children. Move towards this sort of model so we can eliminate rendering
bugs and similar issues.

Additionally, I've been told that the X11 protocol specification's usage
of the word "damage" is incorrect, and that a window is damaged when a
client draws to it; the "damaged region" simply being the region that
contains the pixels the drawing operation touched. Exposures are separate,
and are regions that are generated by the server as a part of inter-window
interactions. Correct this terminology.

For now, don't modify the clients. We'll be removing the ClearDamage request
soon.