server: Add a custom X-WindowTreeChanged event
SubstructureNotify isn't recursive -- it only works on the parent
window.

This mask is selected on the root window, and lets us know when any
change in
the window tree happens, recursively.

The inspector needs to use this to pick up changes in the structure of
windows
in children of the root window. I didn't notice this before because I
wasn't using
the window tree yet!