window-tree: Fix crosshairs incorrectly showing up in certain situations
If you drag the inner window off the right of the parent, and then enter
the child window from the root window, the parent gets an Enter with a
detail "Virtual" it should ignore.

Only respond to the "Inferior" / "Ancestor" details, which are for
direct enters / leaves.

Bug spotted by @Jewelots