Move the kitten that's behind to the bottom
There's a slight amiguity in the text when it says refers to the
window "on top". With this change, it doesn't matter if "top" refers
to the z-axis or the y-axis; they're both the same kitten.

As a bonus, the "L" shape refered to is actually an "L", not an
upside down one :)