server: Remove some complicated processing related to PointerRoot
Keyboard focus has a weird quirk, which is that with the PointerRoot
focus mode, the pointer and keyboard focus are the exact same. This
behavior is part of the server, which allows it to run correctly without
a window manager.

This behavior is, today, considered a misfeature. To simplify the code
of our fake X server, just kill it.