server: Make pixmaps ref-counted
This allows the ServerPixmap for a backing storage for a
ServerWindowDrawTree to be created and destroyed without wrecking the
actual backing storage.

To make ref-counting easier to use from JS, the refcount starts
at *0*. Thus, proper usage of the internal Pixmap is to add an initial
refcount, like with: (new Pixmap()).ref().