server: Add new properties for the "draw tree position"
The old drawing code was implicitly assuming that any draw tree root
would also physically be at 0,0 in the window tree.

Thus, when we went to calculate the bounding region or the window's draw
offset, the draw tree root's position would be added. This worked in the
general case because up until now, the root would be at 0,0, and adding
in 0,0 wouldn't change anything.

To fix this, add new properties for drawTreeX and drawTreeY, which are
the same as normal X/Y if the window is a normal child. However, if the
child is a draw tree root, then these properties give 0,0. Thus, this
gives the correct offset into the pixmap.

At the same time, fix up all the drawing and region calculating code to
correctly use these new properties.