window-tree: Start ignoring NonlinearVirtual again
I misspoke in the commit message of 85e00cf. We want to ignore *all*
Virtual and NonlinearVirtual events. What we'd get in that case is
instead a Nonlinear event. Any event tagged "Virtual" means that we're
zipping past that window on the way to the real thing.

This fixes the case mentioned in 0f0744f in composited-demo.