server: Fix input grab semantics
This is really subtle, but:

* We need to make sure that the winX / winY coordinates map the proper
event window, even when overrwritten by.
* The event window can't be the grabWindow in an owner-events case since
it needs to be reported with respect to the original window.
* We can't throw out events that don't have a proper event window, since
they might be reported with respect to the grab window.

The result is a mess of passing incomplete events around, but is the
best way I can think of to solve the problems. :(